+++ CHANGELOG.md
@@ -18,6 +18,7 @@
 * Updated handling of `command` attributes in `param` tags during parsing
   of XML launch files: Only the output of `stdout` is recorded, and `stderr`
   is now ignored.
+* Added handling of `$(eval ...)` tags in XML launch files.
 
 
 # 1.3.0 (2020-06-02)